Extract step-scheduling graph from composite.py into scheduling.py
composite.py (3523 lines) carried a self-contained step-scheduling subsystem —
find_step_triggers, wire_step_layers, explode_path, merge_collections,
empty_front, find_leaves, build_step_network, build_trigger_state,
find_downstream, determine_steps (+ their lone helper assert_interface). These
operate only on dicts/paths/step-metadata + bigraph_schema.resolve_path, with
zero Composite/Process/Step class coupling, so they move cleanly to
scheduling.py (one-directional: scheduling.py imports nothing from composite).

Pure structural move, no behavior change. composite.py imports the functions
back and re-exports them, so every existing path still resolves:
  - process_bigraph.composite.<fn>  (e.g. genEcoli: from process_bigraph.composite import empty_front)
  - process_bigraph.<fn> top-level  (e.g. vEcoli-composite: from process_bigraph import wire_step_layers)
composite.py: ~3523 -> 3036 lines.

Non-slow suite: 47 passed, 8 skipped (identical to baseline).
